% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corr2cov}
\alias{corr2cov}
\title{From Correlation To Covariance Matrix}
\usage{
corr2cov(corrmat, margprob)
}
\arguments{
\item{corrmat}{The correlation matrix}

\item{margprob}{A list with the marginal probabilities}
}
\value{
The covariance matrix
}
\description{
Converts a correlation matrix for binary variates to a covariance matrix given the
variables marginal probabilities.
}
