% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create2x2Tables.R
\name{create2x2Tables}
\alias{create2x2Tables}
\title{Create 2 x 2 Tables}
\usage{
create2x2Tables(sr)
}
\arguments{
\item{sr}{The output generated by the \code{simulateSRData} function}
}
\value{
A data frame where each row represents a 2 x 2 table. The columns represent: 
  \item{\code{drug_id}}{The ID of the drug}
  \item{\code{event_id}}{The ID of the event}
  \item{\code{prob_drug}}{The marginal probability of that drug}
  \item{\code{prob_event}}{The marginal probability of that event}
  \item{\code{n_corr_drugs}}{The number of drugs that correlate with this drug}
  \item{\code{n_corr_events}}{The number of events that correlate with the given event}
  \item{\code{max_corr_drug}}{The maximum correlation between this drug and other drugs}
  \item{\code{max_corr_event}}{The maximum correlation between this event and other events}
  \item{\code{corr}}{The correlation between the drug-event pair}
  \item{\code{cov}}{The covariance between the drug-event pair}
  \item{\code{commonprob}}{The common probability for the drug-event pair}
  \item{\code{associated}}{\code{TRUE} is there is a non-zero correlation, \code{FALSE} otherwise}
  \item{\code{a}}{Number of times the drug and event appeared together in a report}
  \item{\code{b}}{Number of times the event appeared without the drug in a report}
  \item{\code{c}}{Number of times the drug appeared without the event in a report}
  \item{\code{d}}{Number of times the drug and event both did not appear in a report}
}
\description{
Creates a data frame with all 2 x 2 contigency tables given a simulated data 
set from \code{simulateSRData}. The tables are organized 
as follows: 
\tabular{lccc}{
   \tab event \eqn{j} \tab not event \eqn{j} \tab \emph{total}\cr
  drug \eqn{i} \tab \code{a} \tab \code{c} \tab \code{a} + \code{c}\cr
  not drug \eqn{i} \tab \code{b} \tab \code{d} \tab \code{b} + \code{d}\cr
  \emph{total} \tab \code{a} + \code{b} \tab \code{c} + \code{d} \tab \code{n_reports}
}
}
\examples{
res <- simulateSRData()
tables <- create2x2Tables(res)

}
\seealso{
\code{\link{simulateSRData}}
}
