% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create2x2TablesDAGRcpp}
\alias{create2x2TablesDAGRcpp}
\title{Create 2 x 2 Tables based on DAG generated SR data}
\usage{
create2x2TablesDAGRcpp(reports, betas, prob_drugs, prob_events)
}
\arguments{
\item{reports}{A binary matrix. Each row is a report}

\item{betas}{The matrix with the betas for each pair of variates}

\item{prob_drugs}{A list with the marginal probabilities of the drugs}

\item{prob_events}{A list with the marginal probabilities of the events}

\item{n_parents}{List with the number of parents for each node}
}
\value{
A dataframe. A description of the columns can be found in the commentary
        for the function \code{\link{create2x2TablesDAG}}
}
\description{
Creates a data frame containing all 2 x 2 contingency tables 
from the results generated by generateSRDAG. 
See the R wrapper function \code{\link{create2x2TablesDAG}}
for more information.
}
\seealso{
\code{\link{create2x2TablesDAG}}
}
