% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create2x2TablesRcpp}
\alias{create2x2TablesRcpp}
\title{Create 2 x 2 Tables}
\usage{
create2x2TablesRcpp(reports, corrmat, prob_drugs, prob_events)
}
\arguments{
\item{reports}{A binary matrix. Each row is a report}

\item{corrmat}{The correlation matrix}

\item{prob_drugs}{A list with the marginal probabilities of the drugs}

\item{prob_events}{A list with the marginal probabilities of the events}
}
\value{
A dataframe. A description of the columns can be found in the commentary
        for the function \code{\link{create2x2Tables}}
}
\description{
Creates a data frame containing all 2 x 2 contingency tables 
from the results generated by generateSRData. 
See the R wrapper function \code{\link{create2x2Tables}}
for more information.
}
\seealso{
\code{\link{create2x2Tables}}
}
