% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_beta0.R
\name{f_beta0}
\alias{f_beta0}
\title{Determining the Intercept}
\usage{
f_beta0(beta0, margprob, beta1, margprob_parent)
}
\arguments{
\item{beta0}{The intercept}

\item{margprob}{The marginal probability}

\item{beta1}{The regression coefficient of the parent variable}

\item{margprob_parent}{The marginal probability of the parent}
}
\value{
The absolute difference between the desired marginal probability
and the current marginal probability
}
\description{
\code{f_beta0} is used by \code{\link[=simulateSRS]{simulateSRS()}} to find the appropriate
intercept for a logistic regression model.
}
