% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBlockCorrelationMatrix.R
\name{generateBlockCorrelationMatrix}
\alias{generateBlockCorrelationMatrix}
\title{Block Correlation Matrix}
\usage{
generateBlockCorrelationMatrix(prob_drugs, prob_events, blocksize_drugs,
  rho_drugs, blocksize_events, rho_events, n_correlated_pairs, rho,
  max_iter = 1000, verbose = FALSE)
}
\arguments{
\item{prob_drugs}{List with the marginal probabilities of the drugs}

\item{prob_events}{List with the marginal probabilities of the events}

\item{blocksize_drugs}{The block size for the drugs}

\item{rho_drugs}{The correlation for the drug-drug pairs in the blocks}

\item{blocksize_events}{The block size for the events}

\item{n_correlated_pairs}{The number of drug-event pairs that will have correlation \code{rho}}

\item{rho}{The correlation for the drug-event pairs}

\item{max_iter}{Maximum number of iterations (Default: 1000)}

\item{verbose}{Prints the number of iterations (Default: \code{FALSE})}

\item{rho_drugs}{The correlation for the event-event pairs in the blocks}
}
\value{
\item{corrmat}{The correlation matrix}
        \item{covmat}{The covariance matrix}
        \item{valid}{\code{TRUE} when \code{corrmat} is a correlation matrix, \code{FALSE} otherwise}
}
\description{
Generates a block correlation matrix for the SR data set. It randomly selects 
a number of drug-event, drug-drug, and event-event pairs to exhibit a 
given correlation. It takes the marginal probabilities of the drugs and 
events into account, to make sure that the pairs can indeed show the 
wanted correlation (see the function \code{\link{validCorrelation}}).
\cr\cr 
Since the matrix is generated randomly, it might happen that the resulting matrix
is not a correlation matrix (i.e., it is not positive definite). In that case, 
the process is repeated. One can set the maximal number of iterations with 
the parameter \code{max_iter} (the default is set to 1000).
}
\seealso{
\code{\link{validCorrelation}}
}
