% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateBlockCorrelationMatrixRcpp}
\alias{generateBlockCorrelationMatrixRcpp}
\title{Block Correlation Matrix}
\usage{
generateBlockCorrelationMatrixRcpp(margprob, blocksize, rho)
}
\arguments{
\item{margprob}{List with the marginal probabilities}

\item{blocksize}{The size of the blocks}
}
\value{
A matrix
}
\description{
Block Correlation Matrix
}
\section{Warning}{
 
The matrix that is returned is not guaranteed to be a correlation matrix. 
One still needs to check whether the matrix is indeed postive definite.
}

