% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCorrelationMatrix.R
\name{generateCorrelationMatrix}
\alias{generateCorrelationMatrix}
\title{Correlation Matrix}
\usage{
generateCorrelationMatrix(prob_drugs, prob_events, n_correlated_drugs,
  rho_drugs, n_correlated_events, rho_events, n_correlated_pairs, rho,
  max_iter = 1000, verbose = FALSE)
}
\arguments{
\item{prob_drugs}{List with the marginal probabilities of the drugs}

\item{prob_events}{List with the marginal probabilities of the events}

\item{n_correlated_drugs}{The number of drug-drug pairs that will have correlation \code{rho_drugs}}

\item{rho_drugs}{The correlation for the drug-drug pairs}

\item{n_correlated_events}{The number of event-event pairs that will have correlation \code{rho_events}}

\item{rho_events}{The correlation for the event-event pairs}

\item{n_correlated_pairs}{The number of drug-event pairs that will have correlation \code{rho}}

\item{rho}{The correlation for the drug-event pairs}

\item{max_iter}{Maximum number of iterations (Default: 1000)}

\item{verbose}{Prints the number of iterations (Default: \code{FALSE})}
}
\value{
\item{corrmat}{The correlation matrix}
        \item{valid}{\code{TRUE} when \code{corrmat} is a correlation matrix, \code{FALSE} otherwise}
}
\description{
Generates a correlation matrix for the SR data set. It randomly selects 
a number of drug-event, drug-drug, and event-event pairs to exhibit a 
given correlation. It takes the marginal probabilities of the drugs and 
events into account, to make sure that the pairs can indeed show the 
wanted correlation (see the function \code{\link{validCorrelation}}).
\cr\cr 
Since the matrix is generated randomly, it might happen that the resulting matrix
is not a correlation matrix (i.e., it is not positive definite). In that case, 
the process is repeated. One can set the maximal number of iterations with 
the parameter \code{max_iter} (the default is set to 1000).
}
\seealso{
\code{\link{validCorrelation}}
}
