% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateRandomCorrelationMatrixRcpp}
\alias{generateRandomCorrelationMatrixRcpp}
\title{Random Correlation Matrix}
\usage{
generateRandomCorrelationMatrixRcpp(prob_drugs, prob_events, n_correlated_pairs,
  rho)
}
\arguments{
\item{prob_drugs}{List with the marginal probabilities of the drugs}

\item{prob_events}{List with the marginal probabilities of the events}

\item{n_correlated_pairs}{The number of drug-event pairs that will have correlation \code{rho}}

\item{rho}{The correlation for the drug-event pairs}
}
\value{
\item{corrmat}{A (potential) correlation matrix}
        \item{de_pairs}{A integer matrix with the drug-event pairs}
}
\description{
Generates a correlation matrix for the SR data set. It randomly selects 
a number of drug-event, drug-drug, and event-event pairs to exhibit a 
given correlation. It takes the marginal probabilities of the drugs and 
events into account, to make sure that the pairs can indeed show the 
wanted correlation (see the function \code{\link{validCorrelation}}).
}
\section{Warning}{
 
The matrix that is returned is not guaranteed to be a correlation matrix. 
One still needs to check whether the matrix is indeed postive definite.
}

\seealso{
\code{\link{returnRandomDrugEventPairsRcpp}}, 
         \code{\link{returnRandomDPairsRcpp}},
         \code{\link{validCorrelation}}
}
