% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateReports}
\alias{generateReports}
\title{Generating Reports}
\usage{
generateReports(n_reports, means, L, valid_reports, n_drugs, n_events, verbose)
}
\arguments{
\item{n_reports}{The number of reports generated}

\item{means}{The mean of the multivariate normal distribution}

\item{L}{The Cholesky decomposition of the covariance matrix of the multivariate normal distribution}

\item{valid_reports}{When \code{TRUE}, only valid reports are added}

\item{n_drugs}{The number of drugs}

\item{n_events}{The number of events}

\item{verbose}{Verbosity. If \code{TRUE}, the function prints the 
number of reports that have been processed. Works only when valid_reports is \code{TRUE}}
}
\value{
A logical matrix. Each row is a report. The number of columns is equal to \code{n_drugs + n_events}
}
\description{
Generates a number of reports. First, we sample from a multivariate normal distribution.
We then turn the sampled data into binary data, by checking whether the values are 
larger than zero or not.
\cr\cr
A report normally contains a least one drug and at least one event. In order to make 
sure that is the case, we check whether the generated report is indeed valid 
(see the function \code{\link{validReport}}). One can turn this off by setting
\code{valid_reports} to \code{FALSE}
}
\seealso{
\code{\link{validReport}}
}
