% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateSimpleRandomCorrelationMatrix}
\alias{generateSimpleRandomCorrelationMatrix}
\title{Simple Random Correlation Matrix}
\usage{
generateSimpleRandomCorrelationMatrix(margprob)
}
\arguments{
\item{margprob}{List with the marginal probabilities}
}
\value{
A matrix
}
\description{
Simple Random Correlation Matrix
}
\section{Warning}{
 
The matrix that is returned is not guaranteed to be a correlation matrix. 
One still needs to check whether the matrix is indeed postive definite.
}

