% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnormArma}
\alias{mvrnormArma}
\title{Sampling Multivariate Normal}
\usage{
mvrnormArma(n, mu, L, ncols)
}
\arguments{
\item{n}{The number of samples}

\item{mu}{Vector with the means}

\item{L}{The Cholesky decomposition of the covariance matrix}

\item{ncols}{The number of variates}
}
\value{
A matrix, where each row is a sample
}
\description{
Returns a sample of multivariate normal data. The code is based on the code 
from \link{http://gallery.rcpp.org/articles/simulate-multivariate-normal/}
}
\examples{
arma::mat L = arma::chol(covmat) ;   // covmat is the covariance matrix
arma::mat sample = mvrnormArma(n, means, L, covmat.n_cols) ; 

}
