% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{returnRandomDrugEventPairsRcpp}
\alias{returnRandomDrugEventPairsRcpp}
\title{Random Drug-Event Pairs}
\usage{
returnRandomDrugEventPairsRcpp(prob_drugs, prob_events, n_wanted_pairs, rho)
}
\arguments{
\item{prob_drugs}{List with the marginal probabilities of the drugs}

\item{prob_events}{List with the marginal probabilities of the events}

\item{n_wanted_pairs}{The number of randomly selected pairs}

\item{rho}{The correlation one wants these pairs to have}
}
\value{
A matrix with \code{n_wanted_pairs} rows and two column. Each row are the 
        indices of a drug-event pair (the first index is for the drug, the second
        for the event)
}
\description{
Returns a number of randomly select drug-event pairs that can exhibit a desired
correlation (\code{rho}).
}
\seealso{
\code{\link{validCorrelation}}, \code{\link{sampleRandomRows}}, 
         \code{\link{returnRandomPairsRcpp}},
}
