% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{returnRandomPairsRcpp}
\alias{returnRandomPairsRcpp}
\title{Random Pairs}
\usage{
returnRandomPairsRcpp(margprob, n_wanted_pairs, rho)
}
\arguments{
\item{margprob}{List with the marginal probabilities}

\item{n_wanted_pairs}{The number of randomly selected pairs}

\item{rho}{The correlation one wants these pairs to exhibit}
}
\value{
A matrix with \code{n_wanted_pairs} rows and two column. Each row are the 
        indices of a pair
}
\description{
Returns a number of randomly select (drug-drug or event-event) pairs that can 
exhibit a desired correlation (\code{rho}).
}
\seealso{
\code{\link{validCorrelation}}, \code{\link{sampleRandomRows}}, 
          \code{\link{returnRandomDrugEventPairsRcpp}}
}
