% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleRandomRows}
\alias{sampleRandomRows}
\title{Randows Rows}
\usage{
sampleRandomRows(mat, nrows, n_samples)
}
\arguments{
\item{mat}{A integer matrix}

\item{nrows}{The number of rows of the matrix \code{mat}}

\item{n_samples}{The number of rows one wants to sample}
}
\value{
An integer matrix with \code{n_samples} rows
}
\description{
Returns a random sample of rows from an integer matrix (without replacement).
}
\section{Warning}{

The matrix can contain integers only.
}

