% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDAG.R
\name{simulateDAG}
\alias{simulateDAG}
\title{Spontaneous Reporting Data}
\usage{
simulateDAG(n_reports = 100, n_drugs = 10, n_events = 10,
  alpha_drugs = 1, beta_drugs = 5, alpha_events = 1, beta_events = 5,
  method = "er", exp_degree = 3, theta_drugs = 2,
  n_correlated_pairs = 2, theta = 2, valid_reports = TRUE,
  max_iter = 1000, seed = NULL, verbose = TRUE)
}
\arguments{
\item{n_reports}{Number of reports (Default: 100)}

\item{n_drugs}{Number of drugs (Default: 10)}

\item{n_events}{Number of adverse drug events (Default: 10)}

\item{alpha_drugs}{Alpha parameter for the drug marginal probabilities (Default: 1.0)}

\item{beta_drugs}{Beta parameter for the drug marginal probabilities (Default: 5.0)}

\item{alpha_events}{Alpha parameter for the event marginal probabilities (Default: 1.0)}

\item{beta_events}{Beta parameter for the event marginal probabilities (Default: 5.0)}

\item{n_correlated_pairs}{Number of drug-event pairs that will exhibit a nonzero correlation (Default: 2)}

\item{valid_reports}{If \code{TRUE}, only valid reports (with at least one drug and at least one event) are accepted. (Default: \code{TRUE})}

\item{max_iter}{Maximum number of iterations (Default: 1000)}

\item{seed}{The seed used by the RNG (Default: automatically set)}

\item{verbose}{Verbosity (Default: \code{TRUE})}
}
\value{
\item{sr}{A data frame with the simulated reports. The columns are named \code{drug1}, \code{drug2} ..., \code{event1}, \code{event2}, ...}
        \item{prob_drugs}{A list with marginal probabilities of the drugs}
        \item{prob_events}{A list with marginal probabilities of the events}
        \item{corrmat}{The correlation matrix}
}
\description{
\strong{Valid Reports} Not any binary sequence is a valid report. Each report 
should contain at least one drug and at least one event  (otherwise it would 
never been sent to the spontaneous reporitng sytem). 
While generating reports, we make sure that this is indeed the case. When one does not
want to check the validity and wants to allow any binary sequence, one can set 
\code{valid_reports} to \code{FALSE}.
}
\seealso{
\code{\link{create2x2Tables}}, 
         \code{\link{validReport}}
}
