% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulateReport}
\alias{simulateReport}
\title{Simulate a single Report}
\usage{
simulateReport(n_drugs, n_events, n_parents, beta0, beta1, parent_id)
}
\arguments{
\item{n_drugs}{The number of drugs}

\item{n_events}{The number of events}

\item{n_parents}{A vector with the number of parents for each of the nodes}

\item{beta0}{A vector with the intercepts for the logistic regression models}

\item{beta1}{A vector with the regression coefficients for each node}

\item{parent_id}{A vector with the id of the parent}
}
\value{
A binary vector
}
\description{
Returns a single report for a spontaneous reporting system.
Note that this function does not check whether the report is
'valid', see \code{\link{validReport}}
}
