% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulateReportDAG}
\alias{simulateReportDAG}
\title{Simulate a single report from a DAG}
\usage{
simulateReportDAG(n_drugs, n_events, id, n_parents, max_n_parents, betas,
  verbose)
}
\description{
Simulate a single report from a DAG
}
