% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSRDAG.R
\name{simulateSRDAG}
\alias{simulateSRDAG}
\title{Spontaneous Reporting Data based on a DAG}
\usage{
simulateSRDAG(n_reports = 100, n_drugs = 10, n_events = 10,
  alpha_drugs = 1, beta_drugs = 5, prob_drugs = rbeta(n_drugs,
  alpha_drugs, beta_drugs), alpha_events = 1, beta_events = 5,
  prob_events = rbeta(n_events, alpha_events, beta_events), method = "er",
  exp_degree = 3, theta_drugs = 1.5, n_correlated_pairs = 2, theta = 2,
  seed = NULL, valid_reports = TRUE, verbose = TRUE)
}
\arguments{
\item{n_reports}{Number of reports (Default: 100)}

\item{n_drugs}{Number of drugs (Default: 10)}

\item{n_events}{Number of adverse drug events (Default: 10)}

\item{alpha_drugs}{Alpha parameter for the drug marginal probabilities (Default: 1.0)}

\item{beta_drugs}{Beta parameter for the drug marginal probabilities (Default: 5.0)}

\item{prob_drugs}{Specific array with the marginal probability of the drugs (Default: Beta distribution)}

\item{alpha_events}{Alpha parameter for the event marginal probabilities (Default: 1.0)}

\item{beta_events}{Beta parameter for the event marginal probabilities (Default: 5.0)}

\item{prob_events}{Specific array with the marginal probability of the events (Default: Beta distribution)}

\item{method}{The method used for generating the random DAG for the drugs (Default: \code{'er'}, Erdos-Renyi graph)}

\item{exp_degree}{Average degree between the drugs in the random DAG (Default: 3)}

\item{theta_drugs}{The increase in odds-ratio when there is a connection between two drugs (Default: 1.5)}

\item{n_correlated_pairs}{Number of drug-event pairs that will exhibit a nonzero correlation (Default: 2)}

\item{theta}{Increase in odds-ratio when there is an edge going from a drug to an event (Default: 2.0)}

\item{seed}{The seed used by the RNG (Default: automatically set)}

\item{valid_reports}{If \code{TRUE}, only valid reports (with at least one drug and at least one event) are accepted. (Default: \code{TRUE})}

\item{verbose}{Verbosity (Default: \code{TRUE})}

\item{max_iter}{Maximum number of iterations (Default: 1000)}

\item{DAG}{The DAG used for the simulation. By default generated with the other parameters given to this function}
}
\value{
\item{sr}{A data frame with the simulated reports. The columns are named \code{drug1}, \code{drug2} ..., \code{event1}, \code{event2}, ...}
        \item{prob_drugs}{A list with marginal probabilities of the drugs}
        \item{prob_events}{A list with marginal probabilities of the events}
        \item{adjencency_matrix}{The adjecency matrix of the DAG used}
        \item{nodes}{A tibble with all the information on each node/variate}
}
\description{
\code{simulateSRData} simulates a spontaneous reporting (SR) data set. It allows for setting
the partial probabilities of the drugs and the events, \strong{and} for setting the 
correlations not only between the drugs and the events, but also between the drugs and the 
events themselves. 
\cr\cr
Each report to a SR system contains two lists: 
\enumerate{
  \item the drugs to which the patient was (thought to be) exposed to, and 
  \item the adverse events that the patient experienced. 
}
We will represent each report as a binary vector. The first items represent whether 
the patient was exposed to the drug (\code{1} if he/she was, and \code{0} otherwise). 
The second part represents whether the patient experienced the event or not 
(\code{1} if he/she did, and \code{0} otherwise). For example, if there are 3 drugs and 
4 events in total, a typical report could be 
\deqn{0 1 0 1 1 0 0} 
which represents that the patient was exposed to drug 2 (but not to drug 1 and 3), and 
experienced event 1 and 2 (but not 3 and 4). The simulation results in a binary matrix
where each row is a report. 
\cr\cr
\strong{Valid Reports} Not any binary sequence is a valid report. Each report 
should contain at least one drug and at least one event  (otherwise it would 
never been sent to the spontaneous reporitng sytem). 
While generating reports, we make sure that this is indeed the case. When one does not
want to check the validity and wants to allow any binary sequence, one can set 
\code{valid_reports} to \code{FALSE}.
}
\seealso{
\code{\link{create2x2Tables}}, 
         \code{\link{generateCorrelationMatrix}}, 
         \code{\link{validReport}},
         \code{\link{validCorrelation}}
}
