% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSRData.R
\name{simulateSRData}
\alias{simulateSRData}
\title{Spontaneous Reporting Data}
\usage{
simulateSRData(n_reports = 100, n_drugs = 10, n_events = 10,
  alpha_drugs = 1, beta_drugs = 5, prob_drugs = rbeta(n_drugs,
  alpha_drugs, beta_drugs), alpha_events = 1, beta_events = 5,
  prob_events = rbeta(n_events, alpha_events, beta_events),
  n_correlated_drugs = 5, rho_drugs = 0.25, n_correlated_events = 5,
  rho_events = 0.25, n_correlated_pairs = 2, rho = 0.25,
  valid_reports = TRUE, max_iter = 1000, seed = NULL, verbose = TRUE)
}
\arguments{
\item{n_reports}{Number of reports (Default: 100)}

\item{n_drugs}{Number of drugs (Default: 10)}

\item{n_events}{Number of adverse drug events (Default: 10)}

\item{alpha_drugs}{Alpha parameter for the drug marginal probabilities (Default: 1.0)}

\item{beta_drugs}{Beta parameter for the drug marginal probabilities (Default: 5.0)}

\item{prob_drugs}{Specific array with the marginal probability of the drugs (Default: Beta distribution)}

\item{alpha_events}{Alpha parameter for the event marginal probabilities (Default: 1.0)}

\item{beta_events}{Beta parameter for the event marginal probabilities (Default: 5.0)}

\item{prob_events}{Specific array with the marginal probability of the events (Default: Beta distribution)}

\item{n_correlated_drugs}{Number of drug-pairs that will exhibit a nonzero correlation (Default: 5)}

\item{rho_drugs}{The correlation of the selected drug-pairs (Default: 0.25)}

\item{n_correlated_events}{Number of event-pairs that will exhibit a nonzero correlation (Default: 5)}

\item{rho_events}{The correlation of the selected event-pairs (Default: 0.25)}

\item{n_correlated_pairs}{Number of drug-event pairs that will exhibit a nonzero correlation (Default: 2)}

\item{rho}{The correlation of the selected drug-event pairs (Default: 0.25)}

\item{valid_reports}{If \code{TRUE}, only valid reports (with at least one drug and at least one event) are accepted. (Default: \code{TRUE})}

\item{max_iter}{Maximum number of iterations (Default: 1000)}

\item{seed}{The seed used by the RNG (Default: automatically set)}

\item{verbose}{Verbosity (Default: \code{TRUE})}
}
\value{
\item{sr}{A data frame with the simulated reports. The columns are named \code{drug1}, \code{drug2} ..., \code{event1}, \code{event2}, ...}
        \item{prob_drugs}{A list with marginal probabilities of the drugs}
        \item{prob_events}{A list with marginal probabilities of the events}
        \item{corrmat}{The correlation matrix}
}
\description{
\code{simulateSRData} simulates a spontaneous reporting (SR) data set. It allows for setting
the marginal probabilities of the drugs and the events, \strong{and} for setting the 
correlations not only between the drugs and the events, but also between the drugs and the 
events themselves. 
\cr\cr
Each report to a SR system contains two lists: 
\enumerate{
  \item the drugs to which the patient was (thought to be) exposed to, and 
  \item the adverse events that the patient experienced. 
}
We will represent each report as a binary vector. The first items represent whether 
the patient was exposed to the drug (\code{1} if he/she was, and \code{0} otherwise). 
The second part represents whether the patient experienced the event or not 
(\code{1} if he/she did, and \code{0} otherwise). For example, if there are 3 drugs and 
4 events in total, a typical report could be 
\deqn{0 1 0 1 1 0 0} 
which represents that the patient was exposed to drug 2 (but not to drug 1 and 3), and 
experienced event 1 and 2 (but not 3 and 4). The simulation results in a binary matrix
where each row is a report. 
\cr\cr
\strong{Marginal Probabilities}
The marginal probabilities of the drugs and the events can be set specifically with the 
parameters \code{prob_drugs} and \code{prob_events}. In case they are not set, they are 
randomly sampled from a Beta(\eqn{\alpha}, \eqn{\beta}) distribution. One can set the 
\eqn{\alpha} and \eqn{\beta} parameters for the drugs and the events seperately 
with \code{alpha_drugs}, \code{beta_drugs}, \code{alpha_events} and \code{beta_events}. 
By default, the \eqn{\alpha}'s and the \eqn{\beta}'s are set to 1 and 5, respectively. 
\cr\cr 
\strong{Correlations} One can set the correlations not only between the drugs and 
the events, but also between the drugs and the between events themselves. One needs 
to specify the number of drug-event, drug-drug and event-event pairs that should exhibit a 
nonzero correlation. This is done through the parameters 
\code{n_correlated_pairs}, \code{n_correlated_drugs} and \code{n_correlated_events}. 
The correlations that these pairs should exhibit are set by the parameters 
\code{rho}, \code{rho_drugs} and \code{rho_events}. In other words, the correlation 
matrix will be set up in such a way that exactly \code{n_correlated_drugs} drug-drug pairs 
have a correlation of \code{rho_drugs}, \code{n_correlated_events} event-event pairs 
have a correlation of \code{rho_events}, and \code{n_correlated_pairs} drug-event pairs 
have a correlation of \code{rho}. 
Since the correlation matrix is generated randomly, it might happen that the resulting 
matrix is not a correlation matrix per se (i.e., it is not positive definite). In that case, 
the process is repeated. One can set the maximal number of iterations with 
the parameter \code{max_iter} (the default is set to 1000). 
\cr\cr
\emph{NOTE:} the marginal probabilities of a pair 
restrict the correlation they can exhibit (see for more information 
\code{\link{validCorrelation}}). In case it is impossible to select the desired number of 
pairs with the desired correlation, an error occurs.  
\cr\cr
\strong{Valid Reports} Not any binary sequence is a valid report. Each report 
should contain at least one drug and at least one event  (otherwise it would 
never been sent to the spontaneous reporitng sytem). 
While generating reports, we make sure that this is indeed the case. When one does not
want to check the validity and wants to allow any binary sequence, one can set 
\code{valid_reports} to \code{FALSE}.
}
\examples{
 
# the marginal probabilities of the drugs and the events 
prob_drugs <- c(0.5, 0.2, 0.8, 0.4, 0.2) 
prob_events <- c(0.4, 0.2, 0.9, 0.6, 0.1)  

# generate the SR data
res <- simulateSRData(n_reports = 5, 
                      prob_drugs = prob_drugs, 
                      prob_events = prob_events, 
                      rho_drugs = 0.5, 
                      rho_events = 0.25, 
                      rho = 0.2, 
                      seed = 1)
                      
# the raw reports: 
res$sr 
#> # A tibble: 5 x 10
#> drug1 drug2 drug3 drug4 drug5 event1 event2 event3 event4 event5
#> #<dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#> 1     1     0     1     0     0      1      0      1      0      0
#> 2     1     0     1     1     0      0      0      1      1      0
#> 3     1     0     1     1     1      0      0      1      1      0
#> 4     1     0     1     1     0      0      0      1      1      0
#> 5     1     0     1     0     0      0      0      1      1      0

# the randomly generated correlation matrix
res$corrmat 
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]  1.0  0.5    0  0.5  0.5 0.00 0.00 0.00 0.20  0.20
#> [2,]  0.5  1.0    0  0.5  0.0 0.00 0.00 0.00 0.00  0.00
#> [3,]  0.0  0.0    1  0.0  0.0 0.00 0.00 0.00 0.00  0.00
#> [4,]  0.5  0.5    0  1.0  0.5 0.00 0.00 0.00 0.00  0.00
#> [5,]  0.5  0.0    0  0.5  1.0 0.00 0.00 0.00 0.00  0.00
#> [6,]  0.0  0.0    0  0.0  0.0 1.00 0.25 0.00 0.00  0.25
#> [7,]  0.0  0.0    0  0.0  0.0 0.25 1.00 0.00 0.00  0.25
#> [8,]  0.0  0.0    0  0.0  0.0 0.00 0.00 1.00 0.25  0.00
#> [9,]  0.2  0.0    0  0.0  0.0 0.00 0.00 0.25 1.00  0.25
#>[10,]  0.2  0.0    0  0.0  0.0 0.25 0.25 0.00 0.25  1.00

# create 2x2 contingency tables for this case:
(tables <- create2x2Tables(res))
#> # A tibble: 25 x 16
#>    drug_id event_id prob_drug prob_event n_corr_drugs n_corr_events max_corr_drug max_corr_event  corr
#>      <int>    <int>     <dbl>      <dbl>        <dbl>         <dbl>         <dbl>          <dbl> <dbl>
#> 1       1        1       0.5        0.4            3             2           0.5           0.25     0
#> 2       2        1       0.2        0.4            2             2           0.5           0.25     0
#> 3       3        1       0.8        0.4            0             2           0.0           0.25     0
#> 4       4        1       0.4        0.4            3             2           0.5           0.25     0
#> 5       5        1       0.2        0.4            2             2           0.5           0.25     0
#> 6       1        2       0.5        0.2            3             2           0.5           0.25     0
#> 7       2        2       0.2        0.2            2             2           0.5           0.25     0
#> 8       3        2       0.8        0.2            0             2           0.0           0.25     0
#> 9       4        2       0.4        0.2            3             2           0.5           0.25     0
#>10       5        2       0.2        0.2            2             2           0.5           0.25     0
#># ... with 15 more rows, and 7 more variables: cov <dbl>, commonprob <dbl>, associated <lgl>, a <int>,
#>#   b <int>, c <int>, d <int>

}
\seealso{
\code{\link{create2x2Tables}}, 
         \code{\link{generateCorrelationMatrix}}, 
         \code{\link{validReport}},
         \code{\link{validCorrelation}}
}
