% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRSim-package.R
\docType{package}
\name{SRSim-package}
\alias{SRSim}
\alias{SRSim-package}
\title{SRSim: Spontaneous Reporting Simulator (SRSim)}
\description{
A package for simulating spontaneous reporting data as used in the field of pharmacovigilance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bips-hb/srsim}
  \item Report bugs at \url{https://github.com/bips-hb/srsim/issues}
}

}
\author{
Louis Dijkstra & Marco Garling
}
\keyword{internal}
