% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{SRSim}
\alias{SRSim}
\alias{SRSim-package}
\title{SRSim}
\description{
A package for simulating data from spontaneous reporting systems (SRSs). 
Spontaneous reporting systems are used to
}
\author{
Louis Dijkstra & Marco Garling
}
