% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{threshold}
\alias{threshold}
\title{Larger than Zero}
\usage{
threshold(raw)
}
\arguments{
\item{raw}{The numerical matrix}
}
\value{
Logical matrix
}
\description{
Turns a numerical matrix into a logical matrix, by checking for
every item whether it is (strictly) larger than zero.
}
