% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validCorrelation}
\alias{validCorrelation}
\title{Correlations and Marginal Probabilities}
\usage{
validCorrelation(rho, p_x, p_y)
}
\arguments{
\item{rho}{The correlation (\eqn{\rho})}

\item{p_x}{The probability \eqn{p(X = 1)}}

\item{p_y}{The probability \eqn{p(Y = 1)}}
}
\value{
\code{TRUE} when the two binary variables can exhibit the given correlation, \code{FALSE} otherwise
}
\description{
Returns \code{TRUE} when two binary variables with the given marginal probabilities
\code{p_x} and \code{p_y} can have a correlation of \code{rho}. Otherwise it returns 
\code{FALSE}.
}
\details{
Suppose two binary variables, \eqn{X} and \eqn{Y}, have the marginal probabilities
\eqn{p(X = 1) = p_X} and \eqn{p(Y = 1) = p_Y}. Their common probability is equal to 
\eqn{p(X = 1, Y = 1) = p_{XY}}. Their correlation, \eqn{\rho}, can then be written as
\deqn{\rho = \frac{p_{XY} - p_{X}p_{Y}}{\sqrt{p_{X} (1 - p_{X}) p_{Y} (1 - p(Y))}}}
One can see that not every correlation is admissable given the marginal probabilities. 
In fact, the correlation must lie within the interval
\deqn{max(-\alpha_X \alpha_Y, -1/(\alpha_X \alpha_Y)) \le \rho \le min(\alpha_i/\alpha_j, \alpha_j / \alpha_i)}
where \eqn{\alpha_X = \sqrt{p_X (1 - p_X)}} and \eqn{\alpha_Y = \sqrt{p_Y (1 - p_Y)}}.
These constraints are a result of the Fréchet inequalities.
}
