% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validReport}
\alias{validReport}
\title{Valid Report}
\usage{
validReport(report, n_drugs, n_events)
}
\arguments{
\item{report}{A logical matrix with one row and \code{n_drugs + n_events} columns}

\item{n_drugs}{The total number of drugs}

\item{n_events}{The total number of events}
}
\value{
\code{TRUE} when the report is valid, \code{FALSE} otherwise
}
\description{
Checks whether the given report is 'valid'. A report should contain at least
one drug and at least one event (otherwise it would never been sent to the 
spontaneous reporitng sytem anyway).
}
