% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create2x2TablesDAGRcpp}
\alias{create2x2TablesDAGRcpp}
\title{Create 2 x 2 Tables}
\usage{
create2x2TablesDAGRcpp(
  reports,
  prob_drugs,
  prob_events,
  n_parents,
  parent_id,
  beta1
)
}
\arguments{
\item{reports}{A binary matrix. Each row is a report}

\item{prob_drugs}{A vector with the marginal probabilities of the drugs}

\item{prob_events}{A vector with the marginal probabilities of the events}

\item{n_parents}{A vector with the number of parents for each of the nodes}

\item{parent_id}{A vector with the id of the parent}

\item{beta1}{A vector with the regression coefficients for each node}
}
\value{
A dataframe. A description of the columns can be found in the commentary
        for the function \code{\link{convert2Tables}}
}
\description{
Creates a data frame containing all 2 x 2 contingency tables 
from the results generated by \code{\link{simulateSRS}}. 
See the R wrapper function \code{\link{convert2Tables}}
for more information.
}
\seealso{
\code{\link{convert2Tables}}
}
