% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDAG.R
\name{createDAG}
\alias{createDAG}
\title{Directed Acyclic Graph for SR Data}
\usage{
createDAG(n_drugs = 10, n_events = 10, n_innocent_bystanders = 5,
  n_correlated_pairs = 2)
}
\arguments{
\item{n_drugs}{Number of drugs (Default: 10)}

\item{n_events}{Number of adverse drug events (Default: 10)}

\item{n_innocent_bystanders}{Number of innocent bystanders (Default: 5)}

\item{n_correlated_pairs}{Number of drug-event pairs that will be associated (Default: 2)}
}
\value{
The DAG as an \code{igraph} object
}
\description{
\code{createDAG} creates a directed acyclic graph that can be used
for simulating SR data
}
