% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulateReportDAG}
\alias{simulateReportDAG}
\title{Simulate a single report from a DAG}
\usage{
simulateReportDAG(n_drugs, n_events, beta0, betas, verbose)
}
\arguments{
\item{n_drugs}{The number of drugs in the SR}

\item{n_events}{The number of events in the SR}

\item{id}{A list with the indices of the drugs and the events}
}
\description{
Returns a single report based on a DAG structure.
}
